+++ src/time/future/future_ext.rs
@@ -1,9 +1,9 @@
 use core::future::Future;
 
-use crate::channel::Parker;
-use crate::stream::IntoStream;
+//use crate::channel::Parker;
+//use crate::time::stream::IntoStream;
 
-use super::{Delay, IntoFuture, Park, Timeout};
+use super::{Delay, IntoFuture, Timeout};
 
 /// Extend `Future` with time-based operations.
 pub trait FutureExt: Future {
@@ -77,20 +77,21 @@ pub trait FutureExt: Future {
     {
         Delay::new(self, deadline.into_future())
     }
-
-    /// Suspend or resume execution of a future.
-    ///
-    /// When this method is called the execution of the future will be put into
-    /// a suspended state until the channel returns `Parker::Unpark` or the
-    /// channel's senders are dropped. The underlying future will not be polled
-    /// while the it is paused.
-    fn park<I>(self, interval: I) -> Park<Self, I::IntoStream>
-    where
-        Self: Sized,
-        I: IntoStream<Item = Parker>,
-    {
-        Park::new(self, interval.into_stream())
-    }
+    /* FIXME channels
+        /// Suspend or resume execution of a future.
+        ///
+        /// When this method is called the execution of the future will be put into
+        /// a suspended state until the channel returns `Parker::Unpark` or the
+        /// channel's senders are dropped. The underlying future will not be polled
+        /// while the it is paused.
+        fn park<I>(self, interval: I) -> Park<Self, I::IntoStream>
+        where
+            Self: Sized,
+            I: IntoStream<Item = Parker>,
+        {
+            Park::new(self, interval.into_stream())
+        }
+    */
 }
 
 impl<T> FutureExt for T where T: Future {}
