+++ src/time/future/mod.rs
@@ -41,14 +41,13 @@
 
 mod delay;
 mod future_ext;
-mod into_future;
-mod park;
+//mod park;
 mod relative_future;
 mod timeout;
 
 pub use delay::Delay;
 pub use future_ext::FutureExt;
-pub use into_future::IntoFuture;
-pub use park::Park;
+//pub use park::Park;
 pub use relative_future::Timer;
+pub use std::future::IntoFuture;
 pub use timeout::Timeout;
