+++ src/time/stream/buffer.rs
