+++ src/time/stream/into_stream.rs
