+++ src/time/stream/mod.rs
@@ -5,7 +5,7 @@ mod debounce;
 mod delay;
 mod interval;
 mod into_stream;
-mod park;
+//mod park;
 mod sample;
 mod stream_ext;
 mod throttle;
@@ -16,7 +16,7 @@ pub use debounce::Debounce;
 pub use delay::Delay;
 pub use interval::{interval, Interval};
 pub use into_stream::IntoStream;
-pub use park::Park;
+//pub use park::Park;
 pub use sample::Sample;
 pub use stream_ext::StreamExt;
 pub use throttle::Throttle;
