+++ src/time/stream/sample.rs
