+++ src/time/task/sleep.rs
@@ -2,10 +2,10 @@ use std::future::Future;
 use std::pin::Pin;
 use std::task::{Context, Poll};
 
+use crate::time::Timer as AsyncTimer;
 use pin_project_lite::pin_project;
-use wstd::time::Timer as AsyncTimer;
 
-use crate::future::Timer;
+use crate::time::future::Timer;
 use crate::time::{Duration, Instant};
 
 /// Sleeps for the specified amount of time.
@@ -50,7 +50,7 @@ impl Timer for Sleep {
     /// Resets the timer to be `Instant::now()` + `Duration` into the future.
     fn reset_timer(self: std::pin::Pin<&mut Self>) {
         let mut this = self.project();
-        this.timer.set_after(**this.dur);
+        this.timer.set_after(*this.dur);
         *this.completed = false;
     }
 }
